+++ crates/lib/src/games/theship/mod.rs
@@ -1,0 +1,8 @@
+/// The implementation.
+/// Reference: [server queries](https://developer.valvesoftware.com/wiki/Server_queries)
+pub mod protocol;
+/// All types used by the implementation.
+pub mod types;
+
+pub use protocol::*;
+pub use types::*;