+++ crates/lib/src/protocols/types.rs
@@ -18,6 +18,8 @@ pub enum ProprietaryProtocol {
     Minecraft(Option<minecraft::types::Server>),
     FFOW,
     JC2M,
+    Savage2,
+    Mindustry,
 }
 
 /// Enumeration of all valid protocol types
@@ -41,6 +43,8 @@ pub enum GenericResponse<'a> {
     Valve(&'a valve::Response),
     Unreal2(&'a unreal2::Response),
     #[cfg(feature = "games")]
+    Mindustry(&'a crate::games::mindustry::types::ServerData),
+    #[cfg(feature = "games")]
     Minecraft(minecraft::VersionedResponse<'a>),
     #[cfg(feature = "games")]
     TheShip(&'a crate::games::theship::Response),
@@ -48,6 +52,8 @@ pub enum GenericResponse<'a> {
     FFOW(&'a crate::games::ffow::Response),
     #[cfg(feature = "games")]
     JC2M(&'a crate::games::jc2m::Response),
+    #[cfg(feature = "games")]
+    Savage2(&'a crate::games::savage2::Response),
 }
 
 /// All player types
@@ -228,33 +234,33 @@ impl TimeoutSettings {
 
     /// Get the number of retries if there are timeout settings else fall back
     /// to the default
-    pub const fn get_retries_or_default(timeout_settings: &Option<TimeoutSettings>) -> usize {
+    pub const fn get_retries_or_default(timeout_settings: &Option<Self>) -> usize {
         if let Some(timeout_settings) = timeout_settings {
             timeout_settings.get_retries()
         } else {
-            TimeoutSettings::const_default().get_retries()
+            Self::const_default().get_retries()
         }
     }
 
     /// Get the read and write durations if there are timeout settings else fall
     /// back to the defaults
     pub const fn get_read_and_write_or_defaults(
-        timeout_settings: &Option<TimeoutSettings>,
+        timeout_settings: &Option<Self>,
     ) -> (Option<Duration>, Option<Duration>) {
         if let Some(timeout_settings) = timeout_settings {
             (timeout_settings.get_read(), timeout_settings.get_write())
         } else {
-            let default = TimeoutSettings::const_default();
+            let default = Self::const_default();
             (default.get_read(), default.get_write())
         }
     }
 
     /// Get the connect duration given timeout settings or get the default.
-    pub const fn get_connect_or_default(timeout_settings: &Option<TimeoutSettings>) -> Option<Duration> {
+    pub const fn get_connect_or_default(timeout_settings: &Option<Self>) -> Option<Duration> {
         if let Some(timeout_settings) = timeout_settings {
             timeout_settings.get_connect()
         } else {
-            TimeoutSettings::const_default().get_connect()
+            Self::const_default().get_connect()
         }
     }
 
@@ -337,22 +343,22 @@ impl ExtraRequestSettings {
     }
     /// [Sets protocol
     /// version](ExtraRequestSettings#structfield.protocol_version)
-    pub fn set_protocol_version(mut self, protocol_version: i32) -> Self {
+    pub const fn set_protocol_version(mut self, protocol_version: i32) -> Self {
         self.protocol_version = Some(protocol_version);
         self
     }
     /// [Sets gather players](ExtraRequestSettings#structfield.gather_players)
-    pub fn set_gather_players(mut self, gather_players: bool) -> Self {
+    pub const fn set_gather_players(mut self, gather_players: bool) -> Self {
         self.gather_players = Some(gather_players);
         self
     }
     /// [Sets gather rules](ExtraRequestSettings#structfield.gather_rules)
-    pub fn set_gather_rules(mut self, gather_rules: bool) -> Self {
+    pub const fn set_gather_rules(mut self, gather_rules: bool) -> Self {
         self.gather_rules = Some(gather_rules);
         self
     }
     /// [Sets check app ID](ExtraRequestSettings#structfield.check_app_id)
-    pub fn set_check_app_id(mut self, check_app_id: bool) -> Self {
+    pub const fn set_check_app_id(mut self, check_app_id: bool) -> Self {
         self.check_app_id = Some(check_app_id);
         self
     }